store trivia on tokens
Introduces the necessary steps to extract and store trivia on tokens

The serialization step now also stores the underlying source
and a cursor through the source. When we serialize a token,
we take everything in between the cursor position and the token
start position and extract that as trivia.

To speed up offset computation, we compute the line offsets
for the file beforehand.

We convert AstExprConstantString into a style of token.

Right now, this serializes trivia into one big chunk. And also,
it does not split trivia into leading and trailing trivia.
We will probably want to split trivia into whitespace + comments.

We should follow roslyn syntax rules where everything up to
and including a newline character is treated as trailing trivia
for the previous token, and everything after that is leading trivia
for the next token. This makes code modding nicer, as we can
treat nodes as discrete chunks.