Handle other example (with local definition)
We need to be careful with the order of serialization, as
otherwise it can mess up the cursor and trivia tracking.
This commit reorganises some of the serialization so it is in line

TODO: The serialization of commas is wrong, since we do it
too late. This is going to cause an assertion failure once we
hit an example that is comma separated